+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Update how long text that has a line break or is horizontally aligned in a cell (stubs, labels, table body).
 
+* Update latex to assign styling when multiple stub columns exist (#2103). (@thebioengineer)
+
 # gt 1.2.0
 
 * The new `fmt_number_si()` function format numeric values with SI prefixes and an optional unit (which could be also be obtained from a separate column), automatically selecting the appropriate prefix to keep the mantissa in a readable range (#1999). (#2060)